test: extract webhook secret helper for vitest compatibility
Move requireWebhookSecret out of server-only module so unit tests can
import it without triggering the server-only guard.
